Remove a part of testcondvar.c that does not make sense
The test assumed that if we signal three times. each of the three
workers will wake once. But it is equally well allowed be that one
worker wakes three times and the others not at all.

Interestingly, the assumption generally seems to hold on Windows and
Linux, where this test has been run very often, but it fails on
FreeBSD.