Deal with bind() returning 1.
On current Fedora (37, kernel 6.0.16-300.fc37.x86_64), when binding to
both IPv4 and IPv6, things work a bit differently now then anywhere
else, or so it seems.

What we used to do if we want both IPv4 and IPv6:
- create socket for IPv6
- set option IPV6_V6ONLY to off (0)
- bind either port 0 (let the system figure out a free port) or a specified one
- listen
- create socket for IPv4
- bind the port that was used in the previous step
- listen

If we try that now, the second bind returns 1 (!) if the address that we
try to bind to is "all".  If it is "localhost", things work as before.
Not only does bind return 1, it also gives us a different port (!).

It turns out, if bind does return 1, we can close the socket since the
first bind already bound to both protocols.  You don't see that in the
netstat output, but you can connect using both IPv4 and IPv6.

On the mac, something similar seems to be happening (I didn't check the
bind return value), but netstat reports that the socket is bound to both
IPv4 and 6 (in the "all" case).