add tests for dayname() and monthname() on a timestamp column (besides a date column). The server console generates errors: #2023-05-17 19:21:40: client3: parse_date: ERROR: Syntax error in date. #2023-05-17 19:21:40: client3: VARconvert: ERROR: 22018!conversion of string '%A' to type timestamp failed. #2023-05-17 19:21:40: client3: SQLrun: INFO: Executing: select {fn dayname(ts)} from dtm
#2023-05-17 19:21:40: client3: parse_date: ERROR: Syntax error in date.
#2023-05-17 19:21:40: client3: VARconvert: ERROR: 22018!conversion of string '%B' to type timestamp failed.
#2023-05-17 19:21:40: client3: SQLrun: INFO: Executing: select {fn monthname(ts)} from dtm

This is because dayname() and monthname() are replaced in the parser into date_to_str() which only accepts a date
It would be better if we had functions const char * date_dayofweekname() and const char * date_monthname() in gdk_time.c similar to date_dayofweek() and date_month() but returning a const char * instead of bte.
That would also be faster than the current date_to_str(dt, format) format parser.