Initial implementation of direct masking
The strimps data structure holds an array of 64 bit masks that can be
used to directly construct the bitstring corresponding to an input
string.

Currently this is a 256*256=65536 array of uint64_t (524.288 bytes in
total). The value at index i is 0 unless the pair i is one of the
header pairs. It that case the value is a 64 bit quantity with a bit
turned on at the index of the pair among the header pairs.