Fix an off-by-one bug in strimps iteration
This bug would create different bitstrings for the strings 'Karen'
and 'Karen%'. Although running a like query would not be affected
because "like 'Karen'" would not use strimps filtering, now that we
are developing startswith and related functions, patterns like 'Karen'
use strimps and need to be taken into account.