Remove SELECT privileges check from merge table optimizer
We do not need to check the SELECT privileges of child tables (or their
cols) of a merge table. If a user/role has SELECT privs for the merge
table then they must be able to SELECT data even if they don't have
SELECT privs on the child tables. The same way that someone can SELECT
through a VIEW.