[backported from pp_hashjoin]
initial work on simplified trace output for pipeline execution parts.
For the pipeline blocks we now disable the sqlprofiler (ie normaly saveing each statement with ticks etc in bats)
and collect for all workers and loop iterations the sum of the ticks (beware we use += ie no real atomics)
these are then at the end of the block inserted in the normal trace bats.

So what you get is one line per instruction in the loop. With current counts for arguments
and results (incorrect last iteration on one of the workers). And the sum of the ticks, so
we may divide this by number of workers.