Support hot snapshotting huge bats and log files
The regular tar format does not allow member files >8GiB because the
size field in the per-member header is 11 octal digits plus a NUL byte.
However, because we and most tar implementations don't need the NUL
byte, it actually worked up to 64GiB.

This patch adds support for binary size headers as originally introduced
by GNU tar. The new limit is 8 exabyte which is sufficient for the
foreseeable future.