Normalize URL when creating a REMOTE TABLE
This is also going to require upgrade code.

In the tests, references to node2 are created with the old url syntax,
node3 with the new.

Note that the old syntax mapi:monetdb://localhost:50002/ normalizes to

    monetdb://localhost.:50002/

with a period after the localhost.  This is because in mapi: urls,
'localhost' forces a connection over TCP while in monetdb: urls
'localhost' also allows Unix Domain sockets.