Fix block leak in allocator
fixes #7619.

Function sa_alloc tries to satisfy requests from the "current" block
of memory.  When exhausted, the current block is retired and a newly
allocated block becomes the current block. However, when a large
request comes in, it gets its own private block and the current block
is not retired.  The cutoff point between these two behaviors was a
branch "if (sz > SA_BLOCK) do not retire". This turns out to be very
suboptimal.

The problem is that suballocators request blocks of size SA_BLOCK
from the root allocator.  This means that for every new suballocator,
the root allocator unnecessarily retired the current block.

For example, every prepared statement creates a new suballocator.
This makes the root allocator retire the current block, allocate a
new block for the suballocator and then another new block to replace
its retired block.  The block for the suballocator is eventually
reused when (if!) the suballocator is destroyed but the unnecessarily
retired blocks just accumulate.

By writing "if (sz >= SA_BLOCK)" instead, newly created suballocators
leave the current block of the root allocator in place, wasting much
less memory.