Support compressed ON CLIENT
Syntax: ON '{algo}' CLIENT with algo=lz4/gz/xz/bz2

This makes the server decompress incoming and compress outgoing ON
CLIENT streams.

The client is not aware of any difference, in particular you must
make sure that the client doesn't already decompress the data
or compresses it a second time!