make getFileSize return a int64_t rather than size_t
It's dangerous to convert a file size to size_t on 32 bit platforms.

We use int64_t rather than uint64_t because in some places file sizes
are stored in a lng, which is signed.  Also, st_size is usually a signed
type anyway, so using a signed type means fewer tricky signedness
conversions.